+++ js/main.js
@@ -6,6 +6,6 @@ ZYXer's Basecode Version 46.1.1 (2020-04-18)
 import * as Game from "./core/Game.js";
 
 
-jQuery(document).ready(function() {
+document.addEventListener("DOMContentLoaded", () => {
     Game.start();
 });